\name{meta.plot}
\alias{meta.plotClustersForScatter}
\alias{meta.plotClusters}
\alias{meta.plotScatter}
\alias{meta.plotGate}
\alias{meta.plotGating}
\alias{meta.plotExpResult}
\alias{meta.plotExpClusters}

\encoding{utf8}

\title{
Ploting of immunoClust Meta-clustering Results
}

\description{
Thess functions collect plotting helper routinest of the results 
\code{\link{meta.process}} and relies on the prelimary gating
which is also included with the meta.process function.
}

\usage{
meta.plotClustersForScatter(meta, include, filter=0)

meta.plotClusters( meta, include=c() )

meta.plotScatter(meta)

meta.plotGate(meta, gating, pre, gates, pattern=c(),
png.size=1024, filter=0)

meta.plotGating(meta, pre, pattern=c(), png.size=1024, filter=0)

meta.plotExpResult(meta, exp, pattern=c(), png_pre, plot.ellipse=FALSE, 
plot.class=FALSE, png.size=1024, filter=0, ellipse.quantile=0.95)

meta.plotExpClusters(meta, exp, include=c(), png_pre, png.size=1024, 
plot.ellipse=FALSE, plot.class=FALSE, class.col=c(), filter=0,
ellipse.quantile=0.95, desc=NULL, N=NULL)

}

\arguments{
\item{meta}{The list-object returned by the function meta.process.}
\item{exp}{The vector of \code{immunoClust-objects} used for meta-clustering.}

\item{include}{A numeric vector specifying which clusters will be shown on the
plot.  By default, all clusters are included.} 
\item{pattern}{A character array which give the decription path for the gate to
be plotted.}
\item{gating}{A list object containing the hierarchical gating information as 
generated within the meta-clustering process of immunoCust.}
\item{pre}{File prefix for the output png image files. The name for the
png-file is generated by concatinating \code{pre} with the description of the 
gate.}
\item{png_pre}{Same as \code{pre} (see above).}
\item{png.size}{ Pixel width and height for the output the png image files. }
\item{plot.ellipse}{A logical value indicating whether the cluster 90\% 
percentil boundary is to be drawn or not.}
\item{ellipse.quantile}{ Quantile value for the clusters boundary ellipsis. }
\item{plot.class}{ Indicating whether the plotted cell-clusters are colored by 
the meta-clusters or individually. }
\item{class.col}{ Color(s) of the plotting cell-clusters.  May specify a 
different color for each cluster.}

\item{desc}{ A character vector for the parameter description. }
\item{N}{ An integer for the maximum number of observations to be plotted. By
default all observations are plotted.}
\item{filter}{ A numeric number to filter out meta-clusters which have less
number of cell-clusters. }
\item{gates}{ An optional numeric matrix with the gate positions in each 
parameter. }

}

\value{
Plots the clustering results on an appropriatei plotting device or png-file.  

\describe{
    \item{meta.plotClustersForScatter}{Uses the scatter-clustering results of
    the meta object.}
    \item{meta.plotClusters}{Short-cut for plotting the meta-clustering 
    results.}
    \item{meta.plotScatter}{Short-cut for plotting the scatter-clustering 
    result.}
    \item{meta.plotGate}{Plots the meta-clustering results in one hierachy 
    level and iterates through the sub-levels of the gate.}
    \item{meta.plotGating}{ Plots the meta-clustering gating results in each
    hierarchy level stored in png-files in the directory given by \code{pre}.}
    \item{meta.plotExpResults}{ Plots the the cell-events belonginig to 
    meta-clusters within a gate for each sample in a separate file. The gate 
    is specified by \code{pattern}. } 
    \item{meta.plotExpClusters}{ Plots the cell-events belonging the 
    meta-clusters in \code{include} for each sample used for meta-clustering 
    in a separate file. }
}
}

\author{
Till Sörensen \email{till-antoni.soerensen@charite.de}
}

\seealso{
\code{\link{meta.process}}
}

\examples{
data(dat.meta)
meta.plotClusters(dat.meta)
}
